bugfixing
git-svn-id: https://svn.code.sf.net/p/openuss/code/trunk/openuss-plexus@485 60046477-0c10-0410-bf4b-e552a4ff1768

+++ plexus/plexus-web/src/main/java/org/openuss/web/upload/UploadFileManager.java
@@ -69,7 +69,7 @@ public class UploadFileManager extends BaseBean{
 		Validate.notNull(fileInfo, "Parameter fileInfo must not be null!");
 		
 		File file = files.get(fileInfo);
-		if (!file.delete()) {
+		if (file != null && !file.delete()) {
 			logger.error("couldn't delete tempory file "+ file.getAbsolutePath() +" for "+fileInfo.getName());
 			file.deleteOnExit();
 		}
